+++ packages/@expo/metro-runtime/src/__mocks__/LoadingView.ts
@@ -1,4 +1,0 @@
-export default {
-  showMessage: jest.fn(),
-  hide: jest.fn(),
-};